fix(ai-task): bound the emitted stderr line, not just the retained buffer
consumeStderr trimmed stderrPending after the emit loop, so the 64KB cap
only ever applied to what was left over for the next call: a single
emitted line could reach stderrPendingLimit plus the length of one pipe
read. macOS delivered small chunks and stayed under the cap by luck;
Linux delivered ~24KB at a time and blew past it, which is why
`bounds newline-free stderr` failed on every CI run and passed locally.

Trim the newline-free head before the emit instead.

The regression test used to drive a real pipe, which is what made its
outcome depend on the platform. Split the argument-in/result-out half of
the broker program into a fragment that is spliced into the shipped
source verbatim, and test it standalone with the read size pinned as a
parameter (128B / 4KB / 24KB / 64KB). A composition test asserts the
broker still contains the fragment byte for byte, so the code under test
and the code that ships cannot drift apart.
